requirements.txt
# Bot
aiogram~=3.18.0
# Postgres
SQLAlchemy~=2.0.38
asyncpg~=0.30.0
# Redis
redis~=5.2.1
# Config
pydantic-settings~=2.4.0
# Another
openai
loguru~=0.7.2
APScheduler~=3.11.0
